+++ scratch.js
@@ -22,14 +22,21 @@ const nlp = require('./src/index')
 
 // nlp(`yet soft and yielding like a nerf ball`).debug()
 
-let doc = nlp(`Getting ready for whacking day? What's whacking day?`)
-doc.post('  ', true)
+// let doc = nlp(`Getting ready for whacking day? What's whacking day?`)
+// doc.post('  ', true)
 
-let doc = nlp(`Your dreams may vary from those of Globex Corporation, its subsidiaries and shareholders.`)
-doc.organizations().post('©', true)
+// let doc = nlp(`Your dreams may vary from those of Globex Corporation, its subsidiaries and shareholders.`)
+// doc.organizations().post('©', true)
 
 // doc
 // .match('whacking day')
 // .pre(`'`, true)
 // .post(`'`, true)
-console.log(doc.text())
+// console.log(doc.text())
+
+// console.log(
+//   nlp(`...and my butt smells, and i like to kiss my own butt`)
+//     .clauses()
+//     .find(d => d.has('@hasEllipses'))
+//     .text()
+// )